backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
sqlmodel>=0.0.14
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-multipart>=0.0.7
pydantic>=2.6.0
pydantic-settings>=2.1.0
requests>=2.31.0
httpx>=0.26.0

asyncpg>=0.29.0
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
